+++ jam/asyncio/jwt/lists/json.py
@@ -4,7 +4,7 @@ import datetime
 from asyncio import to_thread
 from typing import Literal
 
-from jam.asyncio.jwt.lists.__abc_lists__ import ABCList
+from jam.asyncio.jwt.lists.__abc_list_repo__ import ABCList
 
 
 try:
