+++ jam/asyncio/jwt/lists/redis.py
@@ -3,7 +3,7 @@
 import datetime
 from typing import Literal
 
-from jam.asyncio.jwt.lists.__abc_lists__ import ABCList
+from jam.asyncio.jwt.lists.__abc_list_repo__ import ABCList
 
 
 try:
